BUG: fixed #615
Former-commit-id: a6038035363993439328c409f8d625c2af455862
Former-commit-id: c8d93c3de8396761124955d9b533497e57040084

+++ scripts/sct_maths.py
@@ -326,7 +326,7 @@ def erode(data, radius):
 def smooth(data, sigmas):
     assert len(data.shape) == len(sigmas)
     from scipy.ndimage.filters import gaussian_filter
-    return gaussian_filter(data, sigmas)
+    return gaussian_filter(data.astype(float), sigmas)
 
 
 def get_data(list_fname):
